refactoring and more typing
